Align GitLab source control adapter structure
